move openapi generator back to librefit-api; define and call 'link' command
