+++ librefit-api/.gitignore
@@ -1,0 +1,8 @@
+wwwroot/*.js
+node_modules
+typings
+dist
+rest/
+rest/.openapi-generator/
+.openapi-generator/
+.cache/
