+++ librefit-api/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "librefit-api",
+  "version": "1.0.0",
+  "description": "Package containing the librefit-api OpenAPI specification",
+  "main": "rest/api.js",
+  "scripts": {
+    "link": "npm link && cd ../librefit-web && npm link ../librefit-api && npm install ../librefit-api --save-dev",
+    "build": "tsc"
+  },
+  "author": "Stefan Poindl",
+  "license": "MIT",
+  "devDependencies": {
+    "@rollup/plugin-commonjs": "^24.0.1",
+    "typescript": "^4.9.5",
+    "axios": "^1.3.3"
+  }
+}
