+++ librefit-web/.gitignore
@@ -8,5 +8,4 @@ node_modules
 !.env.example
 vite.config.js.timestamp-*
 vite.config.ts.timestamp-*
-/src/lib/gen/
 /.cache/
